+++ uns_builder_test.go
@@ -2,32 +2,34 @@ package resolution
 
 import (
 	"net/http"
+	"os"
 	"testing"
 
 	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/stretchr/testify/assert"
 )
 
-func TestUnsBuilder(t *testing.T) {
-	t.Parallel()
-	builder := NewUnsBuilder().SetEthereumNetwork("goerli").SetL2EthereumNetwork("mumbai")
-	_, err := builder.Build()
-	assert.Nil(t, err)
-	assert.NotNil(t, uns.l1Service.contractBackend)
-	assert.NotNil(t, uns.l1Service.metadataClient)
-	assert.NotNil(t, uns.l1Service.supportedKeys)
-	assert.NotNil(t, uns.l1Service.proxyReader)
+func getL1TestProviderUrl() string {
+	if os.Getenv("L1_TEST_NET_RPC_URL") != "" {
+		return os.Getenv("L1_TEST_NET_RPC_URL")
+	}
 
-	assert.NotNil(t, uns.l2Service.contractBackend)
-	assert.NotNil(t, uns.l2Service.metadataClient)
-	assert.NotNil(t, uns.l2Service.supportedKeys)
-	assert.NotNil(t, uns.l2Service.proxyReader)
+	panic("L1_TEST_NET_RPC_URL is not set!")
+}
+
+func getL2TestProviderUrl() string {
+	if os.Getenv("L2_TEST_NET_RPC_URL") != "" {
+		return os.Getenv("L2_TEST_NET_RPC_URL")
+	}
+
+	panic("L2_TEST_NET_RPC_URL is not set!")
 }
 
 func TestUnsBuilderSetBackend(t *testing.T) {
 	t.Parallel()
-	backendl1, _ := ethclient.Dial("https://goerli.infura.io/v3/c5da69dfac9c4d9d96dd232580d4124e")
-	backendl2, _ := ethclient.Dial("https://polygon-mumbai.infura.io/v3/c5da69dfac9c4d9d96dd232580d4124e")
+
+	backendl1, _ := ethclient.Dial(getL1TestProviderUrl())
+	backendl2, _ := ethclient.Dial(getL2TestProviderUrl())
 	builder := NewUnsBuilder().SetEthereumNetwork("goerli").SetL2EthereumNetwork("mumbai")
 	builder.SetContractBackend(backendl1)
 	builder.SetL2ContractBackend(backendl2)
@@ -37,10 +39,38 @@ func TestUnsBuilderSetBackend(t *testing.T) {
 	assert.Equal(t, backendl2, uns.l2Service.contractBackend)
 }
 
+func TestUnsBuilderSetProviderUrl(t *testing.T) {
+	t.Parallel()
+
+	builder := NewUnsBuilder()
+	builder = builder.SetEthereumNetwork("goerli").SetContractBackendProviderUrl(getL1TestProviderUrl())
+	builder = builder.SetL2EthereumNetwork("mumbai").SetL2ContractBackendProviderUrl(getL2TestProviderUrl())
+	uns, err := builder.Build()
+	assert.Nil(t, err)
+	assert.NotNil(t, uns)
+}
+
+func TestUnsBuilderSetProxyBackend(t *testing.T) {
+	t.Parallel()
+
+	builder := NewUnsBuilder().SetUdClient("test")
+	uns, err := builder.Build()
+
+	assert.Nil(t, err)
+	assert.Equal(t, uns.l1Service.networkId, 1)
+	assert.Equal(t, uns.l2Service.networkId, 137)
+}
+
 func TestUnsBuilderSetMetadataClient(t *testing.T) {
 	t.Parallel()
 	client := &http.Client{}
+
+	backendl1, _ := ethclient.Dial(getL1TestProviderUrl())
+	backendl2, _ := ethclient.Dial(getL2TestProviderUrl())
 	builder := NewUnsBuilder().SetEthereumNetwork("goerli").SetL2EthereumNetwork("mumbai")
+	builder.SetContractBackend(backendl1)
+	builder.SetL2ContractBackend(backendl2)
+
 	builder.SetMetadataClient(client)
 	uns, err := builder.Build()
 	assert.Nil(t, err)
@@ -48,10 +78,23 @@ func TestUnsBuilderSetMetadataClient(t *testing.T) {
 	assert.Equal(t, client, uns.l2Service.metadataClient)
 }
 
+func TestUnsBuilderChecksContractBackend(t *testing.T) {
+	t.Parallel()
+
+	var expectedError *UnsConfigurationError
+
+	builder := NewUnsBuilder().SetEthereumNetwork("goerli").SetL2EthereumNetwork("mumbai")
+	uns, err := builder.Build()
+	assert.Nil(t, uns)
+	assert.NotNil(t, err)
+	assert.ErrorAs(t, err, &expectedError)
+}
+
 func TestUnsBuilderChecksL2ContractBackend(t *testing.T) {
 	t.Parallel()
 	var expectedError *UnsConfigurationError
-	backendl1, _ := ethclient.Dial("https://goerli.infura.io/v3/c5da69dfac9c4d9d96dd232580d4124e")
+
+	backendl1, _ := ethclient.Dial(getL1TestProviderUrl())
 	builder := NewUnsBuilder().SetEthereumNetwork("goerli").SetL2EthereumNetwork("mumbai")
 	builder.SetContractBackend(backendl1)
 	_, err := builder.Build()
@@ -62,7 +105,8 @@ func TestUnsBuilderChecksL2ContractBackend(t *testing.T) {
 func TestUnsBuilderChecksL1ContractBackend(t *testing.T) {
 	t.Parallel()
 	var expectedError *UnsConfigurationError
-	backendl2, _ := ethclient.Dial("https://goerli.infura.io/v3/c5da69dfac9c4d9d96dd232580d4124e")
+
+	backendl2, _ := ethclient.Dial(getL2TestProviderUrl())
 	builder := NewUnsBuilder().SetEthereumNetwork("goerli").SetL2EthereumNetwork("mumbai")
 	builder.SetL2ContractBackend(backendl2)
 	_, err := builder.Build()