+++ uns_service.go
@@ -10,10 +10,10 @@ import (
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/vm"
-	kns "github.com/jgimeno/go-namehash"
-	"github.com/unstoppabledomains/resolution-go/v2/dnsrecords"
-	"github.com/unstoppabledomains/resolution-go/v2/namingservice"
-	"github.com/unstoppabledomains/resolution-go/v2/uns/contracts/proxyreader"
+	"github.com/unstoppabledomains/resolution-go/v3/dnsrecords"
+	"github.com/unstoppabledomains/resolution-go/v3/namingservice"
+	"github.com/unstoppabledomains/resolution-go/v3/uns/contracts/proxyreader"
+	"github.com/unstoppabledomains/resolution-go/v3/utils"
 )
 
 // Uns is a naming service handles Unstoppable domains resolution.
@@ -29,6 +29,7 @@ type UnsService struct {
 	Layer                  string
 	networkId              int
 	blockchainProviderUrl  string
+	metadataServiceUrl     string
 }
 
 type MetadataClient interface {
@@ -38,8 +39,8 @@ type MetadataClient interface {
 var unsZeroAddress = common.HexToAddress("0x0")
 
 func domainNameToTokenId(domainName string) *big.Int {
-	normalizedName := normalizeName(domainName)
-	namehash := kns.NameHash(normalizedName)
+	normalizedName := utils.NormalizeName(domainName)
+	namehash := utils.UnsEnsNameHash(normalizedName)
 	return namehash.Big()
 }
 
@@ -62,6 +63,34 @@ func (c *UnsService) data(domainName string, keys []string) (*struct {
 	}
 
 	return &data, nil
+}
+
+func (c *UnsService) getAddress(domainName, family, token string) (string, error) {
+	tokenID := domainNameToTokenId(domainName)
+
+	// verify first if domain exists
+	_, err := c.owner(domainName)
+	if err != nil {
+		return "", err
+	}
+
+	data, err := c.proxyReader.GetAddress(&bind.CallOpts{Pending: false}, family, token, tokenID)
+
+	if err != nil {
+		return "", err
+	}
+
+	return data, nil
+}
+
+func (c *UnsService) reverseOf(addr string) (string, error) {
+	domain, err := c.proxyReader.ReverseNameOf(&bind.CallOpts{Pending: false}, common.HexToAddress(addr))
+
+	if err != nil {
+		return "", err
+	}
+
+	return domain, nil
 }
 
 func (c *UnsService) records(domainName string, keys []string) (map[string]string, error) {
@@ -157,8 +186,8 @@ func (c *UnsService) httpUrl(domainName string) (string, error) {
 func (c *UnsService) locations(domainNames []string) (map[string]namingservice.Location, error) {
 	tokenIDs := make([]*big.Int, 0, len(domainNames))
 	for _, domainName := range domainNames {
-		normalizedName := normalizeName(domainName)
-		namehash := kns.NameHash(normalizedName)
+		normalizedName := utils.NormalizeName(domainName)
+		namehash := utils.UnsEnsNameHash(normalizedName)
 		tokenID := namehash.Big()
 		tokenIDs = append(tokenIDs, tokenID)
 	}
@@ -256,7 +285,7 @@ func (c *UnsService) isSupportedDomain(domainName string) (bool, error) {
 	if extension == "zil" {
 		return false, nil
 	}
-	namehash := kns.NameHash(extension)
+	namehash := utils.UnsEnsNameHash(extension)
 	tokenID := namehash.Big()
 	data, err := c.proxyReader.Exists(&bind.CallOpts{Pending: false}, tokenID)
 	if err != nil {
@@ -266,14 +295,14 @@ func (c *UnsService) isSupportedDomain(domainName string) (bool, error) {
 }
 
 func (c *UnsService) tokenURI(domainName string) (string, error) {
-	normalizedName := normalizeName(domainName)
-	namehash := kns.NameHash(normalizedName)
+	normalizedName := utils.NormalizeName(domainName)
+	namehash := utils.UnsEnsNameHash(normalizedName)
 	return c.tokenUriByNamehash(namehash)
 }
 
 func (c *UnsService) tokenURIMetadata(domainName string) (TokenMetadata, error) {
-	normalizedName := normalizeName(domainName)
-	namehash := kns.NameHash(normalizedName)
+	normalizedName := utils.NormalizeName(domainName)
+	namehash := utils.UnsEnsNameHash(normalizedName)
 	return c.tokenURIMetadataByNamehash(namehash)
 }
 
@@ -316,7 +345,7 @@ func (c *UnsService) unhash(domainHash string) (string, error) {
 }
 
 func (c *UnsService) namehash(domainName string) (string, error) {
-	namehash := kns.NameHash(domainName)
+	namehash := utils.UnsEnsNameHash(domainName)
 	return namehash.String(), nil
 }