+++ zns.go
@@ -3,11 +3,12 @@ package resolution
 import (
 	"encoding/json"
 	"strings"
+	"time"
 
-	"github.com/unstoppabledomains/resolution-go/v2/namingservice"
+	"github.com/unstoppabledomains/resolution-go/v3/namingservice"
 
 	"github.com/Zilliqa/gozilliqa-sdk/provider"
-	"github.com/unstoppabledomains/resolution-go/v2/dnsrecords"
+	"github.com/unstoppabledomains/resolution-go/v3/dnsrecords"
 )
 
 // Zns is a naming service handles .zil domains resolution.
@@ -88,6 +89,10 @@ func (zb *znsBuilder) Build() (*Zns, error) {
 		znsRegistry = znsTestnetRegistry
 	}
 	return &Zns{provider: zb.provider, znsRegistry: znsRegistry}, nil
+}
+
+func (z *Zns) DomainExpiry(domainName string) (time.Time, error) {
+	return time.Time{}, &MethodIsNotSupportedError{NamingServiceName: namingservice.UNS}
 }
 
 // State Get raw data attached to domain.
@@ -287,5 +292,9 @@ func (z *Zns) Namehash(domainName string) (string, error) {
 }
 
 func (z *Zns) Unhash(_ string) (string, error) {
+	return "", &MethodIsNotSupportedError{NamingServiceName: namingservice.ZNS}
+}
+
+func (z *Zns) ReverseOf(_ string) (string, error) {
 	return "", &MethodIsNotSupportedError{NamingServiceName: namingservice.ZNS}
 }